+++ aeron-driver/src/main/c/aeron_min_flow_control.c
@@ -347,8 +347,8 @@ int aeron_tagged_flow_control_strategy_to_string(
     return snprintf(
         buffer,
         buffer_len - 1,
-        "group_tag: %" PRId64 ", group_min_size: %" PRId32 ", receiver_count: %" PRIu32,
+        "group_tag: %" PRId64 ", group_min_size: %" PRId32 ", receiver_count: %" PRIu64,
         strategy_state->group_tag,
         strategy_state->group_min_size,
-        (uint32_t)strategy_state->receivers.length);
+        (uint64_t)strategy_state->receivers.length);
 }
